Refactor ISRs to avoid code duplication
For the very most part, the code of motor1_ISR and motor2_ISR are
exactly the same with only difference in using either HAL_MTR1 or
HAL_MTR2.  Move this code to a separate function that takes motor number
as argument.
By doing so, motor2 is now also torque controlled.