BUGFIX: PosConv was running at wrong rate.
Counter was initialized to 0 but only increased after the first
comparision to the decimation value.  This means, for decimation 1,
posconv was only executed every other time since 0 >= 1 is false.

Fixed by increasing *before* comparision (alternatively the counter
could be initialized with 1, but this seems not so nice to me).

Also added the missing command to reset counter.