Change communication protocoll
It is better to not have data of both m otors together in one frame due
to timing issues (not clear when to send when data comes with different
timing).
Therefore the layout has changed a bit.

Status message was extended by "motor ready" flags and the corresponding
mailbox set to send only one byte (which is currently enough).