Fix & remove code that blocked till CAN message is send
There was a bug that prevented the CAN_send() function from waiting till
TA is set (!= is evaluated before &).  This is fixed by adding
parenthesis.  However, with this fix, the code blocks too long so it
breaks realtime requirements of the ISR.
For now the blocking code is simply disabled as it is not necessary.  It
could be used for error detection, though (e.g. to check before
requesting a new transmission, if the last one is already finished).