Transmit CAN messages in a separate timer ISR
Sending stuff on CAN takes too long to handle it properly in the 10 kHz
motor ISRs. Instead do it in a separate ISR that is called by an CPU
timer only at the rate at which we actually want to send (currently 1
kHz).
This leads to a far more stable communication (no package loss or other
problems in first tests).