Change CAN protocol
Go back to the original idea of sending data of both motors together
(i.e. one frame for current of both motors, one frame for position,
etc.). This makes the data structure more clear. Since a separate timer
is now used for sending frames, the original problem of when to send the
data is resolved.

New structure:

Id    | Mailbox | Description
0x010 | 15      | status
0x020 | 14      | current
0x030 | 13      | position
0x040 | 12      | velocity
0x050 | 11      | ADCINA6/ADCINB6