Add error handling.
When an error occurs, the corresponding flag in gErrors is set (so far
only encoder error is checked).  Errors are generally considered
critical.  If at least one error flag is set, system is disabled and the
red LED is turned on.
As long as there is no error, the red LED is turned off (i.e. system
enabled is only indicated by the green LED from now on).