Set position offset via button on GPIO 26
Assumes that a button is connected to GPIO pin 26 and GND. When the button is
pressed, pin 26 is pulled to low and the "reset zero position offset" flag is
set.  While it is set, the current position offset from zero is stored and
subtracted from positions that are sent via CAN (i.e. the current position
becomes the zero position).
The flag is not cleared automatically but is only dependent on pin 26 (i.e.
position will be zeroed until the button is released).