Use PosConv for index position
In theory using the current PosConv position in the encoder index ISR instead
of the exact position based on posn_index_latch, should be less accurate.  In
practice, however, it turns out that comparing the exact values with other
position values coming from PosConv (e.g. when computing index-to-zero offset)
leads to varying inaccuracies.  This might be due to some adjustment in PosConv
when encoders are aligned?
Anyway the error is negligible small, so using current PosConv position in the
index ISR leads to much better results.